Add MainActivity.{java,smali} and others
The ultimate goal of this project includes getting rid of Java and possibly
other Java-related tools.

The Dalvik bytecode format is available on Google's website, but in order to
have something to start with, a Dalvik bytecode assembler may be better.
The build script requires the smali/baksmali JARs present on the machine.